backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
chromadb==0.5.23
sentence-transformers==3.4.1
google-generativeai==0.8.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
pypdf==5.4.0
